third_party/WebKit/LayoutTests/fast/dom/Document/missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
[
    "adoptNode()",
    "createAttribute()",
    "createAttributeNS('http://www.w3.org/2000/svg')",
    "createAttributeNS()",
    "createCDATASection()",
    "createComment()",
    "createElement()",
    "createElementNS('http://www.w3.org/2000/svg')",
    "createElementNS()",
    "createEvent()",
    "createProcessingInstruction('xml')",
    "createProcessingInstruction()",
    "createTextNode()",
    "elementFromPoint()",
    "elementFromPoint(0)",
    "execCommand()",
    "getElementById()",
    "getElementsByClassName()",
    "getElementsByName()",
    "getElementsByTagName()",
    "getElementsByTagNameNS('http://www.w3.org/2000/svg')",
    "getElementsByTagNameNS()",
    "importNode()",
    "queryCommandEnabled()",
    "queryCommandIndeterm()",
    "queryCommandState()",
    "queryCommandSupported()",
    "queryCommandValue()",
].forEach(function(expr)
{
    shouldThrow("document." + expr);
});
</script>
</body>
</html>
